requirements.txt
# Core Dependencies
openai>=1.0.0
python-dotenv>=1.0.0

# Speech Recognition
openai-whisper>=20230918
SpeechRecognition>=3.10.0
pyaudio>=0.2.13

# Text-to-Speech
elevenlabs>=0.2.0
pyttsx3>=2.90
gTTS>=2.4.0

# Audio Processing
soundfile>=0.12.1
numpy>=1.24.0
librosa>=0.10.0

# Natural Language Processing
langchain>=0.1.0
tiktoken>=0.5.0

# Utilities
requests>=2.31.0
aiohttp>=3.9.0
pydantic>=2.0.0

# Development
pytest>=7.4.0
pytest-asyncio>=0.21.0
black>=23.0.0
flake8>=6.0.0
